requirements3.txt
sanic==22.12.0
Sanic-Cors==2.2.0
tortoise-orm==0.20.0
aredis==1.1.8
aiofiles==23.2.1
aiohttp==3.8.4
schematics==2.1.1
PyYAML==6.0
tzlocal==5.0.1
aiomysql==0.2.0
Cython==3.0.2
zhipuai==1.0.7
dashscope==1.13.5
qianfan==0.0.5
openai==1.7.1
langchain-openai==0.0.2
tiktoken==0.5.2
langchain==0.1.0
python-dotenv==1.0.0


